+++ pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
-log_cli = 1
+log_cli = 0
 log_cli_level = DEBUG
