+++ src/models/api/get_article_statistics/references/__init__.py
@@ -2,14 +2,18 @@ from typing import Dict, List
 
 from pydantic import BaseModel
 
-from src.models.api.get_statistics.references.links import Links
-from src.models.api.get_statistics.references.reference_types import ReferenceTypes
-from src.models.api.reference_statistics import ReferenceStatistics
+from src.models.api.get_article_statistics.reference_statistics import (
+    ReferenceStatistics,
+)
+from src.models.api.get_article_statistics.references.links import Links
+from src.models.api.get_article_statistics.references.reference_types import (
+    ReferenceTypes,
+)
 
 
 class References(BaseModel):
     """The purpose of this class is to model the statistics
-    the user wants from the get_statistics endpoint
+    the user wants from the get_article_statistics endpoint
 
     We use BaseModel to avoid the cache attribute"""
