+++ src/models/api/get_article_statistics/references/content/aggregate/cs1/cite_book_references.py
