+++ src/models/api/get_article_statistics/references/content/aggregate/cs1/cite_journal_wikidata.py
@@ -1,9 +1,11 @@
-from src.models.api.get_statistics.references.content.aggregate.wikidata import Wikidata
+from src.models.api.get_article_statistics.references.content.aggregate.wikidata import (
+    Wikidata,
+)
 
 
 class CiteJournalWikidata(Wikidata):
     """The purpose of this class is to model the statistics
-    the user wants from the get_statistics endpoint
+    the user wants from the get_article_statistics endpoint
 
     We use BaseModel to avoid the cache attribute"""
