+++ src/models/api/get_article_statistics/references/content/aggregate/cs1_references.py
@@ -1,19 +1,19 @@
 from pydantic import BaseModel
 
-from src.models.api.get_statistics.references.content.aggregate.cs1.cite_book_references import (
+from src.models.api.get_article_statistics.references.content.aggregate.cs1.cite_book_references import (
     CiteBookReferences,
 )
-from src.models.api.get_statistics.references.content.aggregate.cs1.cite_journal_references import (
+from src.models.api.get_article_statistics.references.content.aggregate.cs1.cite_journal_references import (
     CiteJournalReferences,
 )
-from src.models.api.get_statistics.references.content.aggregate.cs1.cite_web_references import (
+from src.models.api.get_article_statistics.references.content.aggregate.cs1.cite_web_references import (
     CiteWebReferences,
 )
 
 
 class Cs1References(BaseModel):
     """The purpose of this class is to model the statistics
-    the user wants from the get_statistics endpoint
+    the user wants from the get_article_statistics endpoint
 
     We use BaseModel to avoid the cache attribute"""
