+++ src/models/api/get_article_statistics/references/content/aggregate/wikidata.py
@@ -3,7 +3,7 @@ from pydantic import BaseModel
 
 class Wikidata(BaseModel):
     """The purpose of this class is to model the statistics
-    the user wants from the get_statistics endpoint
+    the user wants from the get_article_statistics endpoint
 
     We use BaseModel to avoid the cache attribute"""
