+++ src/models/api/get_article_statistics/references/links.py
@@ -7,7 +7,7 @@ from src.models.wikimedia.wikipedia.url import WikipediaUrl
 
 class Links(BaseModel):
     """The purpose of this class is to model the statistics
-    the user wants from the get_statistics endpoint
+    the user wants from the get_article_statistics endpoint
 
     We use BaseModel to avoid the cache attribute"""
