+++ src/models/api/get_article_statistics/references/reference_types.py
@@ -1,11 +1,11 @@
 from pydantic import BaseModel
 
-from src.models.api.get_statistics.references.content import ContentReferences
+from src.models.api.get_article_statistics.references.content import ContentReferences
 
 
 class ReferenceTypes(BaseModel):
     """The purpose of this class is to model the statistics
-    the user wants from the get_statistics endpoint
+    the user wants from the get_article_statistics endpoint
 
     We use BaseModel to avoid the cache attribute"""
