+++ src/models/wikimedia/wikipedia/analyzer.py
@@ -2,29 +2,35 @@ import logging
 from typing import Optional
 
 from src import IASandboxWikibase, Wikibase
-from src.models.api.get_statistics.article_statistics import ArticleStatistics
-from src.models.api.get_statistics.references import Links, References, ReferenceTypes
-from src.models.api.get_statistics.references.content import (
+from src.models.api.get_article_statistics.article_statistics import ArticleStatistics
+from src.models.api.get_article_statistics.reference_statistics import (
+    ReferenceStatistics,
+)
+from src.models.api.get_article_statistics.references import (
+    Links,
+    References,
+    ReferenceTypes,
+)
+from src.models.api.get_article_statistics.references.content import (
     AggregateContentReferences,
     CitationReferences,
     ContentReferences,
     GeneralReferences,
 )
-from src.models.api.get_statistics.references.content.aggregate import (
+from src.models.api.get_article_statistics.references.content.aggregate import (
     CiteQReferences,
     Cs1References,
 )
-from src.models.api.get_statistics.references.content.aggregate.cs1.cite_book_references import (
+from src.models.api.get_article_statistics.references.content.aggregate.cs1.cite_book_references import (
     CiteBookReferences,
 )
-from src.models.api.get_statistics.references.content.aggregate.cs1.cite_journal_references import (
+from src.models.api.get_article_statistics.references.content.aggregate.cs1.cite_journal_references import (
     CiteJournalReferences,
 )
-from src.models.api.get_statistics.references.content.aggregate.cs1.cite_web_references import (
+from src.models.api.get_article_statistics.references.content.aggregate.cs1.cite_web_references import (
     CiteWebReferences,
 )
 from src.models.api.job import Job
-from src.models.api.reference_statistics import ReferenceStatistics
 from src.models.exceptions import MissingInformationError
 from src.models.wikimedia.enums import AnalyzerReturn
 from src.models.wikimedia.wikipedia.article import WikipediaArticle
